+++ packages/expo-dev-launcher/android/src/debug/java/expo/modules/devlauncher/compose/screens/HomeScreen.kt
@@ -1,5 +1,6 @@
 package expo.modules.devlauncher.compose.screens
 
+import android.util.Log
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -9,8 +10,13 @@ import androidx.compose.foundation.layout.displayCutoutPadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.systemBarsPadding
+import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.drawBehind
@@ -27,8 +33,8 @@ import com.composables.core.Scrim
 import com.composables.core.rememberDialogState
 import com.composeunstyled.Button
 import expo.modules.devlauncher.R
-import expo.modules.devlauncher.compose.HomeAction
-import expo.modules.devlauncher.compose.HomeState
+import expo.modules.devlauncher.compose.models.HomeAction
+import expo.modules.devlauncher.compose.models.HomeState
 import expo.modules.devlauncher.compose.primitives.Accordion
 import expo.modules.devmenu.compose.primitives.pulseEffect
 import expo.modules.devlauncher.compose.ui.AppHeader
@@ -39,6 +45,7 @@ import expo.modules.devlauncher.compose.ui.SectionHeader
 import expo.modules.devlauncher.compose.ui.ServerUrlInput
 import expo.modules.devlauncher.compose.utils.withIsLast
 import expo.modules.devlauncher.launcher.DevLauncherAppEntry
+import expo.modules.devlauncher.launcher.errors.DevLauncherErrorInstance
 import expo.modules.devmenu.compose.primitives.Divider
 import expo.modules.devmenu.compose.primitives.Heading
 import expo.modules.devmenu.compose.primitives.RoundedSurface
@@ -46,6 +53,14 @@ import expo.modules.devmenu.compose.primitives.RowLayout
 import expo.modules.devmenu.compose.primitives.Spacer
 import expo.modules.devmenu.compose.primitives.Text
 import expo.modules.devmenu.compose.theme.Theme
+import kotlinx.coroutines.delay
+import kotlin.time.Clock
+import kotlin.time.Duration
+import kotlin.time.Duration.Companion.seconds
+import kotlin.time.ExperimentalTime
+import kotlin.time.Instant
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.setValue
 
 @Composable
 fun HowToStartDevelopmentServerDialog(dialogState: DialogState) {
@@ -87,6 +102,30 @@ fun HowToStartDevelopmentServerDialog(dialogState: DialogState) {
 }
 
 @Composable
+fun CrashReport(
+  crashReport: DevLauncherErrorInstance?,
+  onClick: (report: DevLauncherErrorInstance) -> Unit = {}
+) {
+  if (crashReport == null) {
+    return
+  }
+
+  Spacer(Theme.spacing.large)
+
+  RoundedSurface {
+    Button(onClick = {
+      onClick(crashReport)
+    }) {
+      Text(
+        "The last time you tried to open an app the development build crashed. Tap to get more information.",
+        modifier = Modifier.padding(Theme.spacing.medium)
+      )
+    }
+  }
+}
+
+@OptIn(ExperimentalTime::class)
+@Composable
 fun HomeScreen(
   state: HomeState,
   onAction: (HomeAction) -> Unit,
@@ -94,6 +133,7 @@ fun HomeScreen(
 ) {
   val hasPackager = state.runningPackagers.isNotEmpty()
   val dialogState = rememberDialogState(initiallyVisible = false)
+  val scrollState = rememberScrollState()
 
   HowToStartDevelopmentServerDialog(dialogState)
 
@@ -108,8 +148,17 @@ fun HomeScreen(
 
     Column(
       modifier = Modifier
+        .verticalScroll(scrollState)
         .padding(horizontal = Theme.spacing.medium)
     ) {
+      val crashReport = state.crashReport
+      CrashReport(
+        crashReport = crashReport,
+        onClick = {
+          onAction(HomeAction.NavigateToCrashReport(it))
+        }
+      )
+
       Spacer(Theme.spacing.large)
 
       Row {
@@ -166,12 +215,42 @@ fun HomeScreen(
           val infoColor = Theme.colors.status.info
           val defaultColor = Theme.colors.status.default
           val isFetching = state.isFetchingPackagers
+          var isFetchingUIState by remember { mutableStateOf(isFetching) }
+          var fetchStartTime by remember { mutableStateOf<Instant?>(null) }
+
+          LaunchedEffect(isFetching) {
+            Log.e("DevLauncher", "isFetchingPackagers changed: $isFetching")
+            if (isFetching) {
+              isFetchingUIState = true
+              fetchStartTime = Clock.System.now()
+              return@LaunchedEffect
+            }
+
+            if (!isFetchingUIState) {
+              return@LaunchedEffect
+            }
+
+            val startTime = fetchStartTime
+            if (startTime == null) {
+              isFetchingUIState = false
+              return@LaunchedEffect
+            }
+
+            val elapsedTime = startTime - Clock.System.now()
+            val remainingTime = 2.seconds - elapsedTime
+
+            delay(remainingTime)
+
+            if (!state.isFetchingPackagers) {
+              isFetchingUIState = false
+            }
+          }
 
           Button(
             onClick = {
               onAction(HomeAction.RefetchRunningApps)
             },
-            enabled = !isFetching
+            enabled = !isFetchingUIState
           ) {
             RowLayout(
               modifier = Modifier.padding(Theme.spacing.medium),
@@ -186,7 +265,7 @@ fun HomeScreen(
                       )
                     }
                     .then(
-                      if (isFetching) {
+                      if (isFetchingUIState) {
                         Modifier.pulseEffect(
                           initialScale = 0.95f,
                           targetScale = 2f,
@@ -200,7 +279,7 @@ fun HomeScreen(
               }
             ) {
               Text(
-                if (isFetching) {
+                if (isFetchingUIState) {
                   "Searching for development servers..."
                 } else {
                   "Fetch development servers"
@@ -227,9 +306,9 @@ fun HomeScreen(
         }
       }
 
-      Spacer(Theme.spacing.large)
-
       if (state.recentlyOpenedApps.isNotEmpty()) {
+        Spacer(Theme.spacing.large)
+
         Row {
           Spacer(Theme.spacing.small)
 
@@ -238,7 +317,7 @@ fun HomeScreen(
             rightIcon = {
               Row {
                 Button(onClick = {
-                  onAction(HomeAction.ResetRecentlyOpendApps)
+                  onAction(HomeAction.ResetRecentlyOpenedApps)
                 }) {
                   Text(
                     "Reset",
@@ -276,6 +355,8 @@ fun HomeScreen(
           }
         }
       }
+
+      Spacer(Theme.spacing.large)
     }
   }
 }