+++ packages/expo-dev-menu/android/src/release/java/expo/modules/devmenu/DevMenuManager.kt
@@ -13,6 +13,8 @@ import expo.modules.devmenu.api.DevMenuMetroClient
 import expo.modules.manifests.core.Manifest
 import kotlinx.coroutines.CoroutineScope
 
+const val DEV_MENU_TAG = "[disabled] ExpoDevMenu"
+
 private const val DEV_MENU_IS_NOT_AVAILABLE = "DevMenu isn't available in release builds"
 
 object DevMenuManager : DevMenuManagerInterface {
@@ -70,10 +72,6 @@ object DevMenuManager : DevMenuManagerInterface {
   }
 
   fun getMenuPreferences(): Bundle {
-    throw IllegalStateException(DEV_MENU_IS_NOT_AVAILABLE)
-  }
-
-  override fun getMenuHost(): ReactHostWrapper {
     throw IllegalStateException(DEV_MENU_IS_NOT_AVAILABLE)
   }