+++ packages/expo-router/CHANGELOG.md
@@ -34,6 +34,8 @@
 ### 💡 Others
 
 - Create href preview component ([#37335](https://github.com/expo/expo/pull/37335) by [@Ubax](https://github.com/Ubax))
+- Add formsheet warning ([#37982](https://github.com/expo/expo/pull/37982) by [@Ubax](https://github.com/Ubax))
+
 
 ## 5.1.4 - 2025-07-18